chore: Nvm we have access to math now

+++ kweb3-kotlin/src/main/kotlin/dev/klepto/kweb3/kotlin/multicall/contract/QuickCall.kt
@@ -7,7 +7,6 @@ import dev.klepto.kweb3.core.type.EthArray
 import dev.klepto.kweb3.core.type.EthArray.array
 import dev.klepto.kweb3.core.type.EthBytes
 import dev.klepto.kweb3.core.type.EthUint
-import dev.klepto.kweb3.core.type.EthUint.uint256
 import dev.klepto.kweb3.kotlin.multicall.MulticallExecutor
 
 /**
@@ -20,7 +19,7 @@ interface QuickCall : MulticallExecutor {
 
     override suspend fun execute(allowFailure: Boolean, calls: List<MulticallExecutor.Call>): List<EthBytes?> {
         // TODO: Implement proper endpoint resolution.
-        val gasLimit = uint256(client.network.endpoints()[0].gasCap.toLong() / calls.size)
+        val gasLimit = client.network.endpoints()[0].gasCap / calls.size
         val addrs = array(calls.map { it.address })
         val datas = array(calls.map { it.data })
         val result = execute(gasLimit, gasLimit, addrs, datas)
